Добавляет исправления в задание 4
