+++ index.html
@@ -226,7 +226,9 @@
       </div>
     </section>
   </template>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/util.js"></script>
+  <script type="module" src="js/data.js"></script>
+  <script type="module" src="js/main.js"></script>
   <!-- <script src="js/functions.js"></script> -->
 </body>
 </html>
